Lint and rename
