+++ .dockerignore
@@ -1,0 +1,5 @@
+.dockerignore
+.git
+.gitignore
+.prettierrc
+node_modules
