+++ .eslintrc.js
@@ -1,21 +1,20 @@
 module.exports = {
-    "settings": {
-        "react": {
-            "version": "detect"
-        }
+  settings: {
+    react: {
+      version: "detect",
     },
-    "env": {
-        "browser": true,
-        "node": true,
-        "es6": true,
-        "jest": true
-    },
-    "plugins": [
-        "react"
-    ],
-    "extends": [
-        "eslint:recommended",
-        "plugin:react/recommended"
-    ],
-    "parser": "babel-eslint"
+  },
+  env: {
+    browser: true,
+    node: true,
+    es6: true,
+    jest: true,
+  },
+  plugins: ["react", "prettier"],
+  extends: [
+    "eslint:recommended",
+    "plugin:react/recommended",
+    "plugin:prettier/recommended",
+  ],
+  parser: "babel-eslint",
 };
