+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "axios": "^0.21.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "5.2.0",
@@ -38,11 +39,13 @@
     ]
   },
   "devDependencies": {
-    "axios": "^0.21.0",
     "babel-eslint": "^10.1.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.5",
     "eslint-plugin-react": "^7.21.5",
-    "lint": "^0.7.0"
+    "lint": "^0.7.0",
+    "prettier":"2.2.1",
+    "eslint-plugin-prettier":"^3.2.0",
+    "eslint-config-prettier":"^7.0.0"
   }
 }
