+++ src/__fixtures__/axiosHelpersFixtures.js
@@ -1,0 +1,9 @@
+const { testAccessToken } = require("./GoldifySolofixtures");
+
+exports.basicHeaders = function () {
+  return {
+    headers: {
+      Authorization: "Bearer " + testAccessToken,
+    },
+  };
+};
