--- src/__test_utils__/axiosHelpersTestUtils.js
@@ -1,9 +1,0 @@
-const { testAccessToken } = require("./GoldifyExecuteTestUtils");
-
-exports.basicHeaders = function() {
-  return {
-    headers: {
-      Authorization: "Bearer " + testAccessToken,
-    },
-  };
-};
