+++ src/__tests__/GoldifyApp.test.js
@@ -1,13 +1,13 @@
 import React from "react";
-import { render, screen } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import userEvent from '@testing-library/user-event';
-import GoldifyApp from '../js/GoldifyApp';
+import { render, screen } from "@testing-library/react";
+import "@testing-library/jest-dom";
+import userEvent from "@testing-library/user-event";
+import GoldifyApp from "../js/GoldifyApp";
 import { configure } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import reportWebVitals from '../js/utils/reportWebVitals';
+import reportWebVitals from "../js/utils/reportWebVitals";
 
-configure({adapter: new Adapter()});
+configure({ adapter: new Adapter() });
 
 test("Goldify Header exists on main app", () => {
   render(<GoldifyApp />);
@@ -20,7 +20,7 @@ test("Content changes on click of Get Started", () => {
   const homeBody = screen.getByText(/Welcome to Goldify!/i);
   expect(homeBody).toBeInTheDocument();
 
-  const leftClick = { button: 0 }
+  const leftClick = { button: 0 };
   const getStartedButton = screen.getByText("Get Started").closest("a");
   const homeButton = screen.getByText("Home").closest("a");
   expect(getStartedButton).toHaveAttribute("href", "/goldify");
@@ -36,11 +36,9 @@ test("Content changes on click of Get Started", () => {
 test("Run Web Vitals", () => {
   reportWebVitals(console.log);
 
-  reportWebVitals(
-    (metric) => {
-      console.log(metric);
-    }
-  );
+  reportWebVitals((metric) => {
+    console.log(metric);
+  });
 });
 
 test("Load index js script file without error", () => {
