+++ src/css/GoldifySoloPage.css
