+++ src/index.js
@@ -1,14 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './css/index.css';
-import GoldifyApp from './js/GoldifyApp';
-import reportWebVitals from './js/utils/reportWebVitals';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./css/index.css";
+import GoldifyApp from "./js/GoldifyApp";
+import reportWebVitals from "./js/utils/reportWebVitals";
 
 ReactDOM.render(
   <React.StrictMode>
     <GoldifyApp />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
 
 // If you want to start measuring performance in your app, pass a function
