+++ src/js/landing/GoldifyLandingPage.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component } from "react";
 import "../../css/GoldifyLandingPage.css";
 
 class GoldifyLandingPage extends Component {
@@ -7,7 +7,8 @@ class GoldifyLandingPage extends Component {
       <div>
         <h1 className="landingHeader">Welcome to Goldify!</h1>
         <p>
-          To get started, click &quot;Getting Started&quot; in the top right of the page!
+          To get started, click &quot;Getting Started&quot; in the top right of
+          the page!
         </p>
       </div>
     );
